Styling The Log In
Making sure it links to the homepage.

Making it look more distinct from the background, changed some of the
spacing on the elements.

Added Font Awesome icons because they are awesome (just needed script
link)

Some hacky code and lots of inline styles but should be ok for the demo.